+++ srcs/map_adjasment.c
@@ -6,7 +6,7 @@
 /*   By: mrozhnova <mrozhnova@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/22 14:45:24 by mrozhnova         #+#    #+#             */
-/*   Updated: 2022/06/24 00:29:00 by mrozhnova        ###   ########.fr       */
+/*   Updated: 2022/06/24 01:08:45 by mrozhnova        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 static void	adjasment_9(t_map *map)									// segment 9
 {
-	int	i;
-	int	j;
+	uint8_t	i;
+	uint8_t	j;
 
 	i = map->i_start;
 	while (i <= map->i_max)
@@ -35,10 +35,10 @@ static void	adjasment_9(t_map *map)									// segment 9
 
 void adjasment(t_map *map)
 {
-	int	i, j;
+	uint8_t	i, j;
 
-	int	cols_index = map->cols - 1;
-	int	lines_index = map->lines - 1;
+	uint8_t	cols_index = map->cols - 1;
+	uint8_t	lines_index = map->lines - 1;
 
 	map->map_a[0][0] = map->map_g[0][1] + \
 						map->map_g[1][1] + \
